basic site files
Let me know if the files are correct and need changing, it is very basic and some of the links don't work. I assuming the JS will need to be changed quite a bit to work with our backend. I haven't been able to test it with my GAP so please take a look at it.