Servlets! (See Desciption)
We finally cracked how to do servlets, so we added the 'create event'
and 'set user preferences' functionality, although the latter is not yet
complete! This code is the first commit that actually updates the
DataStore!
NOTE:  Does not seem to work on localhost, but works on Google Cloud
when fully deployed! Woohoo!