Manual Merge of Tom's Branch and Main
Copied code from the query that I helped Tom implement.
Put it in the preferences.jsp file.
Added some of Michael's designs.
This is not a formal stage, this is a half stage. It is unorganised.
I pushed because I had a problem with app engine saying that there was a
clash in tranasactions.